working on Week 1 Day 3 Exercises

+++ module-1/03_Expressions/student-lecture/dotnet/Lecture/04_ReturnDoubleOfTwo.cs
@@ -12,9 +12,9 @@ namespace Lecture
          4. This method currently returns an int. Change it so that it returns a double.
             TOPIC: Return Types
          */
-        public int ReturnDoubleOfTwo()
+        public double ReturnDoubleOfTwo()
         {
-            return 2;
+            return 2.0;
         }
     }
 }
